services/web/requirements.txt
bcrypt==4.0.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46
GeoAlchemy2==0.13.1
marshmallow-sqlalchemy==0.28.1
PyJWT==2.6.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
boto3==1.26.57
flask-cors==3.0.10
Pillow==9.4.0
Flask-SocketIO==5.3.2
eventlet==0.33.0
dnspython==2.2.1